nextdjango/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
django-mfa2==3.0
django-otp==1.5.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
fido2==1.1.3
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
PyJWT==2.9.0
pyotp==2.9.0
python-jose==3.4.0
python-u2flib-server==5.0.1
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
